(#589)	code	complete  info reward

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/BaseRewardRepository.java
@@ -38,8 +38,8 @@ public class BaseRewardRepository implements IRewardRepository {
      * @return
      */
     @Override
-    public Observable<Object> rewardsInfo(long news_id, float amount) {
-        return mInfoMainClient.rewardsInfo(news_id,amount)
+    public Observable<Object> rewardsInfo(long news_id, double amount) {
+        return mInfoMainClient.rewardsInfo(news_id, (float) amount)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread());
     }
